Revert "update fork merge"

+++ gdx/src/com/badlogic/gdx/maps/tiled/TiledMapTileLayer.java
@@ -24,8 +24,8 @@ public class TiledMapTileLayer extends MapLayer {
 	private int width;
 	private int height;
 
-	private int tileWidth;
-	private int tileHeight;
+	private float tileWidth;
+	private float tileHeight;
 
 	private Cell[][] cells;
 
@@ -40,12 +40,12 @@ public class TiledMapTileLayer extends MapLayer {
 	}
 
 	/** @return tiles' width in pixels */
-	public int getTileWidth () {
+	public float getTileWidth () {
 		return tileWidth;
 	}
 
 	/** @return tiles' height in pixels */
-	public int getTileHeight () {
+	public float getTileHeight () {
 		return tileHeight;
 	}
